refactor: remove WorkoutView component and related logic
feat: add SavedExerciseCard component for displaying saved exercises

feat: implement SetInputModal for weight and reps input

feat: create UnsavedExerciseCard for managing unsaved exercises

feat: add workout draft management with useWorkoutDraft hook

feat: introduce utility functions for date formatting and volume calculation

fix: update API interfaces to support superset grouping for exercises
