+++ apps/api/drizzle/0003_conscious_titania.sql
@@ -1,0 +1,2 @@
+ALTER TABLE "exercises" ADD COLUMN "group_id" varchar(50);--> statement-breakpoint
+ALTER TABLE "exercises" ADD COLUMN "group_type" varchar(20);
